Scope search query and popular events query for subdomains. Modify PayPal return to redirect to User ticket's page
Why
Event searches (and related activities) within subdomains are expected
to return only information relating to the resources on the subdomain.
However due to the use of `raw sql` in retrieving the search and popular
events query, the subdomain scoping was being bypassed.
Also the return path for PayPal, hitherto redirects to event page and
this isn't very desirable.

This commit solves this by:
* Incorporating Act_as_tentant subdomain logic in the Arel query for search and popular events
* Creating a dummy unauthenticated action for PayPal return path and making it then redirect to the authenticated User tickets' page

Finishes [#108931598]
Finishes [#108931770]
Finishes [#108932162]