Change Capybara driver to PhantomJs and use Puma as test server.
Why?
The tests have been unstable in various CI environments and given
that the integration tests are the ones often affected, the bug might
not be unconnected to the driver used for the test. Also, given that this
instability is not experienced locally, it becomes more likely that
running selenium driver with the integrated CI might not be reliable.
Also the test are taking longer to run

This commit addresses this by:
* Adding an installation of PhantomJs in the CI scripts
* Using Poltergeist gem to enable Capybara use PhantomJs as driver
* Refactoring some test cases
* Using Puma as Capybara test server