Testing of Eventx API endpoints
WHY:
To ensure API endpoints respond with json objects.
To ensure that proper authentication and authorization of user is done before user can access some endpoints.
To ensure that accurate data is supplied at endpoints

This issue is fixed by:
Creating a separate folder for API testing
Writing tests for each endpoints.
Writing tests to check that  content type of response is json
Writing test to check if user is properly authenticated before he is authorized to perform certain actions.
Writing tests to check if event created by API consumer is saved to the DB